refactor: externalize droid session hook
