+++ packages/plugins/agent-droid/scripts/copy-assets.cjs
@@ -1,0 +1,3 @@
+const { cpSync } = require("node:fs");
+
+cpSync("src/assets", "dist/assets", { recursive: true });
